Move DataInserter into e2e utils
